[AI] Refactor timestamp into separate component
